Fit Conditional Extremes model

Fit Conditional Extremes model using marginal fits form evgam.

fit_ce(
  data,
  vars = c("rain", "wind_speed"),
  marg_prob = list(f = list("response ~ name", "~ name"), tau = 0.95, jitter = TRUE),
  cond_prob = 0.9,
  f = list(excess ~ name, ~1),
  split_data = TRUE,
  fit_no_keef = FALSE,
  output_all = FALSE
)

Arguments

data

List of dataframes at each location containing data for each variable.

vars

Variable names for each location.

marg_prob

Can be a list of arguments to quantile_thresh if a variable quantile is desired, or a numeric value for simple quantile thresholding.

cond_prob

Conditional quantile for dependence modelling.

f

Formula for evgam model.

split_data

if data has variables stacked, unstack.

fit_no_keef

If model doesn't fit under Keef constraints, fit without

output_all

Logical argument for whether to return quantiles, evgam and marginal fits.

Value

Object of type mexDependence for each location.